Correct what a dead notify group member actually does, and unbreak pylint
Two things from review, one of them my own mess.

I claimed a missing member fails without a word anywhere. That is wrong.
Copilot said so and I measured it rather than argued: the unread task
exception does reach the event loop, which reports it as "Task exception
was never retrieved" with a traceback naming the action.

What is actually true is narrower. `asyncio.wait` does not propagate, so
the group action reports success and the caller never learns. The failure
surfaces only as a stray unhandled-task traceback, whenever the task
happens to be collected, naming no group and reading like a bug rather
than something you configured. Still worth a repair, but not the story I
wrote. Corrected in the docstring, the issue text and the documentation.

The documentation also said Spook inspects every notify group. It inspects
every legacy YAML one; the entity-based groups belong to
`group_unknown_members`.

Then pylint. CI runs it over all files and found R0801 between the alert
and person test helpers, which I wrote across #1447 to #1449. That means
main is already red: the pylint workflow only runs on pull requests, so
nothing on main has run it since May.

I missed it by linting changed files, which is what the local hook does.
Duplicate-code is cross-file by nature, so a changed-files run can never
see it. The shared part now lives in `tests/repair_helpers.py` and both
test files keep a thin wrapper for their own setup.

While rewiring those files I also managed to delete two tests, the pair
proving the #1448 behaviour. Restored from main, and the test inventory of
both files now matches main exactly.
