+++ tests/ectoplasms/person/repairs/test_unknown_device_trackers.py
@@ -7,7 +7,6 @@ from __future__ import annotations
 from types import SimpleNamespace
 from typing import TYPE_CHECKING
 
-from homeassistant.const import EVENT_STATE_CHANGED
 from homeassistant.core import State
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers.entity_component import DATA_INSTANCES
@@ -20,6 +19,7 @@ from custom_components.spook.repairs import (
     PersonUnknownDeviceTrackerFixFlow,
     async_create_fix_flow,
 )
+from tests.repair_helpers import async_count_scheduled_inspections
 
 if TYPE_CHECKING:
     from homeassistant.core import HomeAssistant, ServiceCall
@@ -194,38 +194,6 @@ async def test_fix_flow_remove_yaml_person_aborts(
     assert result["reason"] == "not_editable"
 
 
-async def _count_scheduled_inspections(
-    hass: HomeAssistant,
-    entity_id: str,
-    old_state: State | None,
-    new_state: State | None,
-) -> int:
-    """Return how many inspections one state change schedules."""
-    repair = SpookRepair(hass)
-    await repair.async_activate()
-    repair.inspect_debouncer.async_shutdown()
-    calls = 0
-
-    def async_schedule_call() -> None:
-        """Capture scheduled inspections."""
-        nonlocal calls
-        calls += 1
-
-    repair.inspect_debouncer = SimpleNamespace(
-        async_schedule_call=async_schedule_call,
-        async_shutdown=lambda: None,
-    )
-
-    hass.bus.async_fire(
-        EVENT_STATE_CHANGED,
-        {"entity_id": entity_id, "old_state": old_state, "new_state": new_state},
-    )
-    await hass.async_block_till_done()
-
-    await repair.async_deactivate()
-    return calls
-
-
 async def test_state_only_tracker_going_away_is_reported(
     hass: HomeAssistant,
     issue_registry: ir.IssueRegistry,
@@ -263,6 +231,21 @@ async def test_issue_clears_when_a_state_only_tracker_returns(
     await repair._async_inspect_with_cleanup()
 
     assert issue_registry.async_get_issue(DOMAIN, _ISSUE_ID) is None
+
+
+async def _count_scheduled_inspections(
+    hass: HomeAssistant,
+    entity_id: str,
+    old_state: State | None,
+    new_state: State | None,
+) -> int:
+    """Return how many inspections one state change schedules."""
+    repair = SpookRepair(hass)
+    await repair.async_activate()
+
+    return await async_count_scheduled_inspections(
+        hass, repair, entity_id, old_state, new_state
+    )
 
 
 async def test_state_only_tracker_addition_rechecks_person_repairs(
